IS-1685: Use system token when preloading cache
syfooversiktsrv preloades cache from a cronjob so the cache is ready
before veiledere start working, therefore it uses a system token.
syfobehandlendeenhet has a specific endpoint for system token,
but it gets the same info as the obo token endpoint,
so the data filled in cache should be correct regardless.
Set token to nullable in clients and check if token is null in
AzureAdClient to decide whether to use system token or obo token,
that way existing calls to clients remain the same as before.