IS-1831: Get enhet from norg with GT
Only get enhet from syfobehandlendeenhet if GT is UTLAND or UDEFINERT.
Add GT with type BYDEL to pipresponse generator, so we can switch
syfobehandlendeenhet mock with norg mock in tests.
Keep existing preload methods for now, but we can cache api calls to
norg for GeografiskTilknytning we get when preloading pdl cache.